requirements.txt
glob2==0.6
matplotlib==2.0.2
numpy==1.19.1
pandas==1.0.1
Pillow==5.0.0
scikit-learn==0.20.2
scipy==1.2.1
torch==1.2.0
torchvision==0.4.0a0+6b959ee
opencv-python==3.4.0.12